ракета выстреливает, потом детектит, находит - ускорятеся и атакует

+++ Space/Assets/Sripts/Ammunitions/RocketDetectEnemy.cs
@@ -8,7 +8,6 @@ public class RocketDetectEnemy : MonoBehaviour
 {
 
     private Collider2D[] _colliders = new Collider2D[10];
-    // Start is called before the first frame update
     void Start()
     {
         //Detect();
@@ -24,6 +23,7 @@ public class RocketDetectEnemy : MonoBehaviour
     {
         Array.Clear(_colliders, 0, _colliders.Length);
         var count = Physics2D.OverlapCircleNonAlloc(transform.position, 10f, _colliders, LayerMask.GetMask("Enemy"));
+        if (count == 0) return null;
         //Debug.Log($"Count: {count}");
         var list = _colliders.ToList();
         list.RemoveRange(count, list.Count-count);
